# src/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SecureVault is a secure file storage service with S3 integration, user authentication, file sharing, and administrative features.",
        "title": "SecureVault API",
        "termsOfService": "http://localhost:8080/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080/support",
            "email": "support@securevault.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/public/files/owner/{owner_id}": {
            "get": {
                "description": "Get paginated list of all public files for a specific owner (no authentication required)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get all public files by owner ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner ID (UUID)",
                        "name": "owner_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page (max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of public files",
                        "schema": {
                            "$ref": "#/definitions/src_api.PublicFilesListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid owner ID",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/public/files/share/{token}": {
            "get": {
                "description": "Get file details using share token (no authentication required)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get public file details by share token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File share token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File details",
                        "schema": {
                            "$ref": "#/definitions/src_api.PublicFileResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "File is private",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Share link expired",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/public/files/share/{token}/download": {
            "get": {
                "description": "Download file content using share token (no authentication required)",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Download file by share token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File share token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content with appropriate headers",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid token format",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Share link expired or inactive",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Download failed",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/public/files/{id}": {
            "get": {
                "description": "Get file details for public files (no authentication required)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get public file details by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File details",
                        "schema": {
                            "$ref": "#/definitions/src_api.PublicFileResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid file ID",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "File is private",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/public/files/{id}/download": {
            "get": {
                "description": "Download file content for public files (no authentication required)",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Download public file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content with appropriate headers",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid file ID format",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "File not public",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Download failed",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/public/folders/share/{token}": {
            "get": {
                "description": "Get complete folder tree structure with all subfolders and files using share token (no authentication required)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get public folder tree by share token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder share token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder tree with subfolders and files",
                        "schema": {
                            "$ref": "#/definitions/src_api.PublicFolderTreeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Folder not found",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Share link expired",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/src_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "src_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/src_api.ErrorDetail"
                }
            }
        },
...
        "src_api.PublicFileResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "folder_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "original_filename": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "src_api.PublicFilesListResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/src_api.PublicFileInfo"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/src_api.PaginationMeta"
                }
            }
        },
...
        "src_api.PublicFolderTreeResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/src_api.PublicFileInfo"
                    }
                },
                "folder": {
                    "$ref": "#/definitions/src_api.PublicFolderInfo"
                },
                "subfolders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/src_api.PublicFolderTreeResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}